separate event detail organizers into container and template
